+++ Assets/Script/TnTCountDown.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using Unity.Entities;
-
+using UnityEngine.UI;
 public struct TnTCountDown : IComponentData
 {
-    public float countDown;
+    public int countDown;
 }
